fa qualcosa, dovrebbe funzionare XD

+++ header/hashTable.h
@@ -11,7 +11,7 @@
 
 typedef struct node{
     unsigned char str[4];
-    int cursorPos;
+    unsigned long cursorPos;
     unsigned int hash;
     struct node *next;
 }Node;
@@ -21,11 +21,11 @@ typedef struct table{
     Node *list;
 }Table;
 
-Node * createNode(char str[4], int cursorPos, Node *next, Table *table);
+Node * createNode(char str[4], unsigned long cursorPos, Node *next, Table *table);
 Table *createTable(int size);
 void insert(Node *node, Table *table);
 unsigned int hash(unsigned char str[4], int tableSize);
-int searchAndUpdateMatch(Node *node, Table *table);
+unsigned long searchAndUpdateMatch(Node *node, Table *table);
 void clearTable(Table *table);
 
 #endif //SNAPPY_HASHTABLE_H
